refactor: split SQLite store from ChainSyncProtocol
