+++ Cargo.toml
@@ -21,7 +21,6 @@ hex = "0.4.2"
 net2 = "0.2.35"
 regex = "1.4.2"
 reqwest = { version = "0.10.8", features = ["blocking"] }
-rusqlite = { version = "0.24.1", features = ["bundled"] }
 serde = { version = "1.0.117", features = ["derive"] }
 serde_cbor = "0.11.1"
 serde_json = "1.0.59"
@@ -30,7 +29,12 @@ log = "0.4.11"
 [dev-dependencies]
 simple_logger = "1.11.0"
 futures = "0.3.8"
+rusqlite = { version = "0.24.1", features = ["bundled"] }
 
 [[example]]
 name = "common"
+crate-type = ["staticlib"]
+
+[[example]]
+name = "sqlite"
 crate-type = ["staticlib"]
